requirements.txt
streamlit
bertopic
umap-learn
hdbscan
sentence-transformers
transformers
torch
numpy
scikit-learn
pandas
plotly
matplotlib
pyyaml
openpyxl
xlsxwriter
requests
python-dotenv
rich
markdown-it-py
mdurl
pygments
altair
altgraph
attrs
blinker
bottle
bottle-websocket
cachetools
certifi
cffi
charset-normalizer
click
cmdstanpy
colorama
contourpy
cycler
Eel
et_xmlfile
fonttools
future
gevent
gevent-websocket
gitdb
GitPython
greenlet
holidays
idna
importlib_resources
Jinja2
jsonschema
jsonschema-specifications
kiwisolver
MarkupSafe
narwhals
packaging
pefile
pillow
prophet
protobuf
pyarrow
pycparser
pydeck
pyparsing
python-dateutil
pytz
pywin32-ctypes
referencing
rpds-py
six
smmap
stanio
tenacity
toml
tornado
tqdm
typing_extensions
tzdata
urllib3
watchdog
zope.event
zope.interface
pywin32
msal